+++ application/models/Transactions.php
@@ -53,6 +53,33 @@ class Transactions extends MY_Model2
         return $resultset;
     }
 
+    public function getSalesTransactions($stock){
+
+        $this->db->select('*');
+        $this->db->from('transactions t');
+        $this->db->where('t.Stock', $stock);
+        $query = $this->db->get();
+
+        $item = [
+            "DateTime" => "N/A",
+            "Player" => "N/A",
+            "Stock" => "N/A",
+            "Trans" => "N/A",
+            "Quantity" => "N/A"
+        ];
+
+        $resultset = array();
+
+        array_push($resultset, $item);
+
+        if($query->num_rows() != 0)
+        {
+            $resultset = $query->result_array();
+        }
+
+        return $resultset;
+    }
+
     public function getCurrentHoldings($player){
         $resultset = null;